feat: move OutputPlanner to Refitter.Core with primitive-parameter API
